% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_asgmts.R
\name{archive_asgmts}
\alias{archive_asgmts}
\alias{unarchive_asgmts}
\title{Archive or unarchive assignments on the server}
\usage{
archive_asgmts(ids = NULL, server = "", user = "", password = "")

unarchive_asgmts(ids = NULL, server = "", user = "", password = "")
}
\arguments{
\item{ids}{Assignment IDs to archive or unarchive. This accepts a vector so multiple
assignments can be archived or unarchived at once.}

\item{server}{Prefix for the survey server. It is whatever comes before
mysurvey.solutions: [prefix].mysurvey.solutions.}

\item{user}{Username for the API user on the server.}

\item{password}{Password for the API user on the server.}
}
\value{
Assignments corresponding to the ID number provided are archived or
unarchived.
}
\description{
\code{archive_asgmts} archives the assignments corresponding to the ID number(s)
provided.
\code{unarchive_asgmts} unarchive the assignments corresponding to the ID number(s)
provided.
}
\examples{
## archive assignments with ID numbers 4, 18, 195
\dontrun{
archive_asgmts(ids = c(4, 18, 195), server = "lfs2018",
user = "APIuser2018", password = "SafePassword123")
}
## unarchive assignments with ID numbers 4, 18, 195
\dontrun{
unarchive_asgmts(ids = c(4, 18, 195), server = "lf2018",
user = "APIuser2018", password = "SafePassword123")
}
}
