% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interviewers.R
\name{get_interviewers}
\alias{get_interviewers}
\title{Get information about interviewers on the server}
\usage{
get_interviewers(super_names = NULL, super_ids = NULL, server, user, password)
}
\arguments{
\item{super_names}{Usernames of the supervisors that you would like to get the list of
interviewers for. This accepts a vector of multiple supervisor usernames. To get data on
every interviewer account, leave \code{super_names} and \code{super_ids} null.}

\item{super_ids}{User Id of the supervisors that you would like to get the list of
interviewers for. This accepts a vector of multiple supervisor Ids. User Ids can have
dashes or not.}

\item{server}{Prefix for the survey server. It is whatever comes before
mysurvey.solutions: [prefix].mysurvey.solutions.}

\item{user}{Username for the API user on the server.}

\item{password}{Password for the API user on the server.}
}
\value{
A data frame  data frame containining information about interviewer accounts.
}
\description{
\code{get_interviewers} returns a data frame containing information about interviewer
account. Due to the current limitations of the API, only interviewers that are
\strong{not archived/locked} appear in the output data frame. If a supervisor has no
interviewers or all the interviewers on their team are locked/archived, then the output
data frame will contain a row for that supervisor with both InterName and InterId
being NA.
}
\details{
The data frame will have the following columns:
\itemize{
\item \emph{InterName}: Username of interviewer
\item \emph{InterId}: Unique ID of interviewer
\item \emph{SuperName}: Username of interviewer's supervisor
\item \emph{SuperId}: Unique ID of interviewer's supervisor
\item \emph{IsLocked}: If interviewer is currently locked/archived
\item \emph{CreationDate}: Date the interviewer account was created
\item \emph{DeviceId}: Unique Id of the tablet that the interviewer was using on the last
synchronisation to server. If the user has never synced any data with the server
using a tablet, this variable will be null.
}
}
\examples{
\dontrun{
## Get data on all interviewers that are currently not archived on server
all_ints <- get_interviewers(server = "lfs2018",
user = "APIuser2018", password = "SafePassword123")

## Using the usernames for supervisors
supervisors <- c("North_Supervisor", "East_Supervisor")

interviewers <- get_interviewers(super_names = supervisors,
server = "lfs2018", user = "APIuser2018", password = "SafePassword123")

## Using user Ids for supervisors
supervisors <- c("59e60ce7-a6f9-4b0b-a5b3-ab6e8ce76464",
"60d28f7f-6195-4136-a8e8-a0c118218f0d")

interviewers <- get_interviewers(super_ids = supervisors,
server = "lfs2018", user = "APIuser2018", password = "SafePassword123")
}
}
