% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qx.R
\name{get_qx}
\alias{get_qx}
\title{Get all templates imported on the server}
\usage{
get_qx(server, user, password)
}
\arguments{
\item{server}{Prefix for the survey server. It is whatever comes before
mysurvey.solutions: [prefix].mysurvey.solutions.}

\item{user}{Username for the API user on the server.}

\item{password}{Password for the API user on the server.}
}
\value{
A data frame with information about the
imported questionnaires on the server.
}
\description{
\code{get_qx} returns a data frame with all the questionnaires that are currently
imported on the server
}
\examples{
\dontrun{
get_qx(server = "lfs2018", user = "APIuser2018", password = "SafePassword123")
}
}
