% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supers.R
\name{get_supers}
\alias{get_supers}
\title{Get information about supervisors on the server}
\usage{
get_supers(server, user, password)
}
\arguments{
\item{server}{Prefix for the survey server. It is whatever comes before
mysurvey.solutions: [prefix].mysurvey.solutions.}

\item{user}{Username for the API user on the server.}

\item{password}{Password for the API user on the server.}
}
\value{
A data frame  data frame containining information about supervisor accounts.
}
\description{
\code{get_supers} Returns a data frame containing information about supervisors
account. Due to the current limitations of the API, only supervisors that are
\strong{not archived/locked} appear in the output data frame.
The data frame will have the following columns:
\itemize{
\item SuperName: User name of supervisors
\item SuperId: Unique ID of supervisors
\item IsLocked: If supervisor is currently locked/archived
\item CreationDate: Date the supervisor account was created
\item DeviceId: Unique Id of the tablet that the user was using on the last
synchronisation to server. If the user has never synced any data with the server
using a tablet, this variable will be null.
}
}
\examples{
\dontrun{
get_supers(server = "lfs2018", user = "APIuser2018", password = "SafePassword123")
}
}
