% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_stats.R
\name{interview_stats}
\alias{interview_stats}
\title{Get information about an interview}
\usage{
interview_stats(ids = NULL, server = "", user = "", password = "")
}
\arguments{
\item{ids}{Interview Id. The interview Id can have dashes or not.
This accepts a vector of interview Ids. In exported data, this will be the variable
\emph{interview__id}.}

\item{server}{Prefix for the survey server. It is whatever comes before
mysurvey.solutions: [prefix].mysurvey.solutions.}

\item{user}{Username for the API user on the server.}

\item{password}{Password for the API user on the server.}
}
\value{
A data frame with statistics on the interviews associated with
the interview IDs provided.
}
\description{
\code{interview_stats} gets information about an interview, including:
\itemize{
  \item Interview key
  \item Number of unanswered questions
  \item Number of answered questions
  \item Number of questions with errors
  \item Number of questions with comments
  \item Current status of interview
  \item Number of rejections by HQ
  \item Number of rejections by Supervisor
  \item Assignment Id that the interview was generated from
}
}
\examples{
\dontrun{
## Vector of IDs to get stats for
int_ids <- c("5970be83582840df972082c578816dba", "47748cf7aa0641a692a1fffc508a435c",
"61ef820e5bac466e829355abe42bc137", "63053caafa424a8d993cc22be2c9baf0")

## Get statistics for those interviews
interview_stats(ids = int_ids, server = "lfs2018",
user = "APIuser2018", password = "SafePassword123")
}
}
