% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_one.R
\name{start_export}
\alias{start_export}
\title{Start export of the data for a questionnaire}
\usage{
start_export(
  qx_name = NULL,
  template_id = NULL,
  version = NULL,
  export_type = "tabular",
  return_time = FALSE,
  server = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{qx_name}{Name of questionnaire. This is case sensitive and should match the name
of your questionnaire exactly. Do not specify both template_id and qx_name.}

\item{template_id}{Questionnaire Id for the template. Do not include any dashes.
Do not specify both template_id and qx_name.}

\item{version}{Version number of questionnaire}

\item{export_type}{Data type to export. Valid options are:  tablular, stata, spss,
binary, paradata. Default is tabular.}

\item{return_time}{Option to return the time that the export was started. Start time will
be in UTC.}

\item{server}{Prefix for the survey server. It is whatever comes before
mysurvey.solutions: [prefix].mysurvey.solutions.}

\item{user}{Username for the API user on the server.}

\item{password}{Password for the API user on the server.}
}
\value{
Starts the export for the questionnaire.
}
\description{
\code{start_export} starts the export of a version of a questionnaire for a specified
data type
}
\examples{
\dontrun{
start_time <- start_export(qx_name="Labour Force Survey Q1",
version=4, start_time=TRUE,
server = "lfs2018", user = "APIuser2018", password = "SafePassword123")
}
}
